+++ Source/SelfService/Web/microservice/Simple.tsx
@@ -4,10 +4,11 @@ import React from 'react';
 import { Stack } from '@fluentui/react/lib/Stack';
 import { Label } from '@fluentui/react/lib/Label';
 import { TextField } from '@fluentui/react/lib/TextField';
-import { MicroserviceSimple, createMicroservice } from '../store';
+import { createMicroservice } from '../store';
+import { MicroserviceSimple } from '../api/index';
 import { PrimaryButton } from '@fluentui/react/lib/Button';
 import { Guid } from '@dolittle/rudiments';
-import { HttpResponseApplications2 } from '../api';
+import { HttpResponseApplications2 } from '../api/api';
 
 const stackTokens = { childrenGap: 15 };
 
@@ -68,7 +69,6 @@ export const Microservice: React.FunctionComponent<Props> = (props) => {
         ms.extra.ingress.path = ingressPath;
         // TODO land the name we want here
         ms.extra.ingress.domainPrefix = ingressDomainPrefix;
-        ms.extra.ingress.secretNamePrefix = ingressDomainPrefix;
         console.log('onSave', ms);
         createMicroservice(ms.kind, ms);
     };
