+++ Source/SelfService/Web/screens/backupScreen.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 
-import { HttpResponseApplications2 } from '../api';
+import { HttpResponseApplications2 } from '../api/api';
 import { ViewCard } from '../backup/viewCard';
 
 type Props = {
